requirements.txt
streamlit
folium
plotly.express
requests
langchain
langchain_core
langchain_community
pyspark
numpy
pandas
seaborn
matplotlib
joblib
streamlit_folium
scikit-learn
catboost
lightgbm
imblearn
venn_abers
joblib